+++ crates/types-common/src/lib.rs
@@ -1,0 +1,24 @@
+// SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
+//
+// SPDX-License-Identifier: EUPL-1.2
+
+//! Common data types for OpenTalk.
+//!
+//! This crate contains data types that are commonly used in the OpenTalk !
+//! APIs.
+
+#![deny(
+    bad_style,
+    missing_debug_implementations,
+    missing_docs,
+    overflowing_literals,
+    patterns_in_fns_without_body,
+    trivial_casts,
+    trivial_numeric_casts,
+    unsafe_code,
+    unused,
+    unused_extern_crates,
+    unused_import_braces,
+    unused_qualifications,
+    unused_results
+)]